+++ NEWS.md
@@ -5,6 +5,7 @@
 * Python: added `arkimet.Summary.read_yaml` (#230)
 * Python: added `arkimet.Summary.read_json` (#230)
 * Python: added `annotate: bool = False` to `arkimet.Summary.write` (#230)
+* Python: added `annotate: bool = False` to `arkimet.Summary.write_short` (#230)
 
 # New in version 1.27